@@ -46,10 +46,10 @@ object Svg2Compose {
                 val previousGroup = groupStack.peekOrNull()
 
                 // if there is no previous group, this is the root dir, and the group name should be the accessorName
-                val groupName = if(previousGroup == null) accessorName else file.name.toKotlinPropertyName()
-                val groupPackage = previousGroup?.let { group -> "${group.groupPackage}.${group.groupName.second.toLowerCase()}" }
+                val groupName = if (previousGroup == null) accessorName else file.name.toKotlinPropertyName()
+                val groupPackage = previousGroup?.let { group -> "${group.groupPackage}.${group.groupName.second.lowercase(Locale.getDefault())}" }
                     ?: "$applicationIconPackage"
-                val iconsPackage = "$groupPackage.${groupName.toLowerCase()}"
+                val iconsPackage = "$groupPackage.${groupName.lowercase(Locale.getDefault())}"
 
                 val (groupFileSpec, groupClassName) = IconGroupGenerator(
                     groupPackage,
@@ -58,7 +58,7 @@ object Svg2Compose {
 
 
                 val generatedIconsMemberNames: Map<VectorFile, MemberName> =
-                    if(dirIcons.isNotEmpty()) {
+                    if (dirIcons.isNotEmpty()) {
                         val drawables: List<Pair<File, File>> = when (type) {
                             VectorType.SVG -> dirIcons.map {
                                 val iconName = nameRelative(it).withoutExtension
@@ -70,6 +70,7 @@ object Svg2Compose {
 
                                 it to parsedFile
                             }.toList()
+
                             VectorType.DRAWABLE -> dirIcons.toList().map { it to it }
                         }
 
@@ -80,7 +81,7 @@ object Svg2Compose {
                                     groupName
                                 ),
                                 drawableFile.name,
-                                drawableFile.readText()
+                                drawableFile.readText(),
                             )
                         }
 
@@ -88,6 +89,7 @@ object Svg2Compose {
                             icons.values,
                             groupClassName,
                             iconsPackage,
+                            generatePreview
                         )
 
                         val memberNames = writer.generateTo(outputSourceDirectory) { true }
@@ -108,7 +110,7 @@ object Svg2Compose {
                     childGroups = emptyList()
                 )
 
-                if(previousGroup != null) {
+                if (previousGroup != null) {
                     groupStack.pop()
                     groupStack.push(previousGroup.copy(childGroups = previousGroup.childGroups + result))
                 }
@@ -118,7 +120,7 @@ object Svg2Compose {
                 true
             }
             .onLeave {
-                val group = if(groupStack.size > 1)
+                val group = if (groupStack.size > 1)
                     groupStack.pop()
                 else
                     groupStack.peek()
